Move Features into Settings\Features
* This way we don't have an identifier that is both a class name and a
namespace. This actually caused issues doing this rename, where my IDE
incorrectly changed a few usages of things in the `Features` folder.
