+++ src/ImageArchive/Rest.php
@@ -4,7 +4,6 @@ namespace P4\MasterTheme\ImageArchive;
 
 use Exception;
 use P4\MasterTheme\Capability;
-use P4\MasterTheme\Features;
 use P4\MasterTheme\Exception\RemoteCallFailed;
 use P4\MasterTheme\Features\ImageArchive;
 use WP_Http;
